NDRS-453 - move get_finalized_deploys into storage.rs
